refactor: move useContractExistsAtAddress hook to hook folder
